Feedback continued

+++ Server/controllers/user.controller.js
@@ -5,31 +5,36 @@ const bcrypt = require('bcrypt');
 const mongoF = require("../middleware/mongoFunctions");
 
 
-
+/* This function is used for registering a user
+   fullname, email and password are required in request's body
+*/
 module.exports.register = (req, res, next) => {
-    try {
-        const user = new User({
-            fullName: req.body.fullName,
-            email: req.body.email,
-            password: req.body.password
-        });
-        user.save((err, doc) => {
-            if (!err)
-                res.status(201).json({ success: true, message: 'user added' });
-            else {
-                if (err.code == 11000)
-                    res.status(422).send(['Email address already exists']);
-                else
-                    return next(err)
-            }
-        });
-    } catch {
-        return res.status(500).json({
-            message: "failed"
-        });
+        try {
+            const user = new User({
+                fullName: req.body.fullName,
+                email: req.body.email,
+                password: req.body.password
+            });
+            user.save((err, doc) => {
+                if (!err)
+                    return res.status(201).json({ success: true, message: 'user added' });
+                else {
+                    if (err.code == 11000)
+                        return res.status(422).json({ success: false, message: 'Email Already Exists' });
+                    else
+                        return next(err)
+                }
+            });
+        } catch {
+            return res.status(500).json({
+                success: false,
+                message: "failed"
+            });
+        }
     }
-}
-
+    /* This function is used for login a user
+       email and password are required in request's body
+    */
 module.exports.login = async(req, res, next) => {
     try {
         let fetchedUser;
@@ -39,6 +44,7 @@ module.exports.login = async(req, res, next) => {
             const passwordValidated = await bcrypt.compareSync(req.body.password, user.password);
             if (!passwordValidated) {
                 return res.status(401).json({
+                    success: false,
                     message: "Auth failed"
                 });
             } else {
@@ -53,17 +59,22 @@ module.exports.login = async(req, res, next) => {
             }
         } else {
             return res.status(401).json({
+                success: false,
                 message: "Auth failed"
             });
         }
     } catch (err) {
         return res.status(500).json({
+            success: false,
             message: "failed"
         });
     }
 }
 
-
+/* This function is used for creating a notification
+   header, body and type are required in request's body
+   user's ID is extracted by decoding the token from header
+*/
 module.exports.createNotification = async(req, res, next) => {
     try {
         const user = await mongoF.findOneMongoObject(User, { _id: req.userData.userId });
@@ -93,7 +104,9 @@ module.exports.createNotification = async(req, res, next) => {
 
 }
 
-
+/* This function is used for retrieving all notifications of a user
+   user's ID is extracted by decoding the token from header
+*/
 module.exports.getNotifications = async(req, res, next) => {
     try {
         const notifications = await mongoF.findMultipleMongoObjetcs(Notification, { userId: req.userData.userId }, { 'created_at': -1 });
@@ -105,35 +118,47 @@ module.exports.getNotifications = async(req, res, next) => {
             });
         } else {
             return res.status(500).json({
+                success: false,
                 message: "failed"
             });
         }
     } catch {
         return res.status(500).json({
+            success: false,
             message: "failed"
         });
     }
 }
 
-
+/* This function is used for deleting a notifications of a user
+   The notification id is a query parameter
+   user's ID is extracted by decoding the token from header
+*/
 module.exports.deleteNotification =
     (req, res, next) => {
         try {
-            Notification.deleteOne({ _id: req.params.id }).then(result => {
+            Notification.deleteOne({ _id: req.params.id, userId: req.userData.userId }).then(result => {
                 if (result.n > 0) {
                     res.status(200).json({ message: "Deleted successful!" });
                 } else {
-                    res.status(401).json({ message: "Not authorized to delete!" });
+                    res.status(401).json({
+                        success: false,
+                        message: "Not authorized to delete!"
+                    });
                 }
             });
         } catch {
             return res.status(500).json({
+                success: false,
                 message: "failed"
             });
         }
     }
 
-
+/* This function is used for editing a notifications of a user
+   The notification id is a query parameter
+   user's ID is extracted by decoding the token from header
+*/
 module.exports.updateNotification = async(req, res, next) => {
     try {
         const notificationObj = await mongoF.findOneMongoObject(Notification, { _id: req.params.id });
@@ -144,11 +169,14 @@ module.exports.updateNotification = async(req, res, next) => {
                 body: req.body.body,
                 type: req.body.type
             });
-            const updated = await Notification.updateOne({ _id: notification._id }, notification);
+            const updated = await Notification.updateOne({ _id: notification._id, userId: req.userData.userId }, notification);
             if (updated.nModified > 0) {
                 return res.status(201).json({ success: true, message: 'Notification Updated' });
             } else {
-                res.status(401).json({ message: "Unable to update!" });
+                res.status(401).json({
+                    success: false,
+                    message: "Unable to update!"
+                });
             }
         } else {
 
@@ -156,6 +184,7 @@ module.exports.updateNotification = async(req, res, next) => {
         }
     } catch {
         return res.status(500).json({
+            success: false,
             message: "failed"
         });
     }
